+++ lib/ua_inspector/config.ex
@@ -35,13 +35,16 @@ defmodule UAInspector.Config do
   end
 
   @doc """
-  Returns the configured database path or `nil`.
+  Returns the configured database path.
+
+  If the path is not defined the `priv` dir of `:ua_inspector`
+  as returned by `Application.app_dir(:ua_inspector, "priv")` will be used.
   """
-  @spec database_path() :: String.t() | nil
+  @spec database_path() :: String.t()
   def database_path do
     case get(:database_path) do
-      nil -> nil
-      path -> Path.expand(path)
+      nil -> Application.app_dir(:ua_inspector, "priv")
+      path -> path
     end
   end