+++ lib/ua_inspector/storage/server.ex
@@ -43,13 +43,7 @@ defmodule UAInspector.Storage.Server do
         old_ets_tid = ETS.fetch_data(@ets_lookup_table_name, @ets_data_table_name)
         new_ets_tid = ETS.create_data(@ets_data_table_name)
 
-        _ =
-          if Config.database_path() do
-            :ok = do_reload(new_ets_tid)
-          else
-            _ = Logger.warn("Reload error: no database path configured")
-          end
-
+        :ok = do_reload(new_ets_tid)
         :ok = schedule_data_cleanup(old_ets_tid)
         :ok = ETS.update_data(@ets_lookup_table_name, @ets_data_table_name, new_ets_tid)